scripts/rag/requirements.txt
pinecone
sentence-transformers
python-dotenv
requests
beautifulsoup4
